Extract performance horizon row assembly
